requirements.txt
django==2.2
djangorestframework==3.9.2
gunicorn==20.0.4